+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All the following depend on Core Branding v4+.
 
 - `Icon`[BREAKING] `<i>` is for italic, not for icons. Only `<span class="..."></span>` is allowed. [WCAG 2.1 AA - Identify purpose](https://www.w3.org/TR/WCAG21/#identify-purpose)
 - `Button` Add contrast warning for negative toggle. [WCAG 2.1 AA - Contrast](https://www.w3.org/TR/WCAG21/#contrast-minimum)
+- `Tag`/`TagListItem` Add extra WVAG information for `aria-label`.
 
 ## [Unreleased]
 - `ButtonGroup` Added option for vertical button group (depends on core branding 3.2)